# api/swagger.yaml
openapi: 3.0.3
info:
  title: Restic Monitor API
  description: |
    API for monitoring Restic backup repositories. 
    Provides endpoints to check backup status, view snapshots, manage locked repositories, and prune old backups.
  version: 1.0.0
  contact:
    name: Restic Monitor
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://your-server.com/api/v1
    description: Production server
security:
  - basicAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication with username and password
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication using API token
  schemas:
    BackupStatus:
      type: object
      properties:
        name:
          type: string
          description: Unique name of the backup target
          example: "home"
        latestBackup:
          type: string
          format: date-time
          description: Timestamp of the most recent snapshot
          example: "2025-11-22T10:30:00Z"
        latestSnapshotID:
          type: string
          description: ID of the latest snapshot
          example: "a1b2c3d4e5f6g7h8i9j0"
        snapshotCount:
          type: integer
          description: Total number of snapshots in the repository
          example: 42
        fileCount:
          type: integer
          description: Number of files in the latest snapshot
          example: 15234
        health:
          type: boolean
          description: Whether the repository is healthy
          example: true
        statusMessage:
          type: string
          description: Error or status message if health is false
          example: "repository is locked"
        checkedAt:
          type: string
          format: date-time
          description: Timestamp when the status was last checked
          example: "2025-11-22T11:00:00Z"
        disabled:
          type: boolean
          description: Whether monitoring is disabled for this target
          example: false
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          description: Status message
          example: "success"
...
paths:
  /status:
    get:
      summary: Get status of all backup targets
      description: Returns the current status of all configured backup targets including health, snapshot counts, and last check time
      operationId: getStatus
      tags:
        - Status
      responses:
        '200':
          description: Successful response with array of backup statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BackupStatus'
        '401':
          description: Unauthorized - authentication required
          content:
            text/plain:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
...
  /toggle/{name}:
    post:
      summary: Toggle monitoring state for a target
      description: Enables or disables monitoring for the specified backup target
      operationId: toggleTargetDisabled
      tags:
        - Management
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the backup target to toggle
          schema:
            type: string
          example: "home"
      responses:
        '200':
          description: Target state toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: "toggled"
        '400':
          description: Bad request - target name missing
        '401':
          description: Unauthorized
        '404':
          description: Target not found
        '405':
          description: Method not allowed - only POST accepted
        '500':
          description: Toggle operation failed
...